Added prettier and tailwind support
Installed the `Tailwind CSS IntelliSense` extension to VSCode which provide class name suggestions
and linting for your CSS classes.

Also added the `prettier-plugin-tailwindcss` package that provides formatting support for prettier.
This plugin is referenced in the `.prettierrc.cjs` config and helps prettier to order the
tailwind classes in the same order as they appear in the generated CSS file.